% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_pairs_limits.R
\name{bayes_pairs_limits}
\alias{bayes_pairs_limits}
\title{Plot pairs posteriors}
\usage{
bayes_pairs_limits(posteriors, limits)
}
\arguments{
\item{posteriors}{a data.frame with parameters as columns and samples as rows}

\item{limits}{a data.frame with paramaters as columns and two rows, one with upper bounds and one with lower bounds (the order does not matter)}
}
\value{
a gtable with rug plots on the top and lefthand side, parameter names on the diagonal and 2D density plots in other positions
}
\description{
Plot pairs posteriors
}
