% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_pairs_priors.R
\name{bayes_pairs_priors}
\alias{bayes_pairs_priors}
\title{Plot pairs posteriors}
\usage{
bayes_pairs_priors(posteriors, priors)
}
\arguments{
\item{posteriors}{a data.frame with each posterior parameter as a column}

\item{priors}{a data.frame sampled from the prior distribution used to produce the posteriors}
}
\value{
a gtable with rug plots on the top and lefthand side, parameter names on the diagonal and 2D density plots in other positions
}
\description{
Plot pairs posteriors
}
